Move coconut validator api req out of coconut interface
and expose a new cosmos-address endpoint
