+++ validator-api/src/coconut/client.rs
@@ -3,10 +3,11 @@
 
 use crate::coconut::error::Result;
 use multisig_contract_common::msg::ProposalResponse;
-use validator_client::nymd::{Fee, TxResponse};
+use validator_client::nymd::{AccountId, Fee, TxResponse};
 
 #[async_trait]
 pub trait Client {
+    async fn address(&self) -> AccountId;
     async fn get_tx(&self, tx_hash: &str) -> Result<TxResponse>;
     async fn get_proposal(&self, proposal_id: u64) -> Result<ProposalResponse>;
     async fn propose_release_funds(
