+++ validator-api/validator-api-requests/src/lib.rs
@@ -1,4 +1,5 @@
 // Copyright 2022 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod coconut;
 pub mod models;
